Release v1.5
NEW:      Support for backup and restore of prepared queries
UPDATED:  Updated 0.7.5 consul API package
UPDATED:  Updated to 2.0.4 minio-go S3 API package

NOTE:     This and all future releases will be accompanied by a gpg
signed SHA256SUM file.  Details are contained in the README.md in the
root of the repository.